consider empty corpus a non-starter
